requirements.txt
numpy
pandas
python-dotenv
scikit_learn
setuptools
nltk
python-dotenv
matplotlib
seaborn
mlflow
dagshub
xgboost
skorch
fastapi
unicorn
